backend/aws-server/image/requirements.txt
torch==2.2.1
torchvision==0.17.1
Pillow==10.1.0
onnxruntime==1.16.0
opencv-python-headless==4.9.0.80
numpy==1.26.2
